fixed: Osm entities should reference actual primitives of osm layer
The OSM entities were built with a reference to the primitives in the
dataset of the osm download. This created issues in the geometry update.
After this fix, the osm entities refer to the primitives in the dataset
of the osm layer. The dataset of the osm download is cleared after the
downloaded primitives are merged into the osm layer.